+++ src/panels/config/ha-panel-config.ts
@@ -20,7 +20,6 @@ import {
   mdiPuzzle,
   mdiRobot,
   mdiScriptText,
-  mdiServer,
   mdiShape,
   mdiSofa,
   mdiTools,
@@ -256,14 +255,6 @@ export const configSections: { [name: string]: PageNavigation[] } = {
   ],
   general: [
     {
-      component: "server_control",
-      path: "/config/server_control",
-      translationKey: "ui.panel.config.server_control.caption",
-      iconPath: mdiServer,
-      iconColor: "#4A5963",
-      core: true,
-    },
-    {
       path: "/config/updates",
       translationKey: "ui.panel.config.updates.caption",
       iconPath: mdiUpdate,
@@ -444,10 +435,6 @@ class HaPanelConfig extends HassRouterPage {
       helpers: {
         tag: "ha-config-helpers",
         load: () => import("./helpers/ha-config-helpers"),
-      },
-      server_control: {
-        tag: "ha-config-server-control",
-        load: () => import("./server_control/ha-config-server-control"),
       },
       storage: {
         tag: "ha-config-section-storage",
